python-app-demo/vm2-quart/requirements.txt
quart>=0.19
hypercorn>=0.16
opentelemetry-distro>=0.45b0
opentelemetry-exporter-otlp-proto-http>=1.24
opentelemetry-instrumentation-asgi>=0.45b0
opentelemetry-instrumentation-sqlite3>=0.45b0
